requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
python-dotenv==1.0.1
scikit-learn==1.4.0
numpy==1.26.4
nltk==3.8.1
openai==1.12.0
